App/requirements.txt
openai>=0.27.0
python-dotenv>=0.19.2
requests>=2.26.0
